requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.1
openpyxl==3.1.2
python-docx==0.8.11
matplotlib==3.7.1
colorama==0.4.6
Pillow==9.4.0